Refactor TerrainMapFileLoader to BinaryLoaderGameMapLoader and extract GameMapLoader interface
