+++ src/client/components/Maps.ts
@@ -1,7 +1,7 @@
 import { LitElement, css, html } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
+import { terrainMapFileLoader } from "../../core/game/BinaryLoaderGameMapLoader";
 import { GameMapType } from "../../core/game/Game";
-import { terrainMapFileLoader } from "../../core/game/TerrainMapFileLoader";
 import { translateText } from "../Utils";
 
 // Add map descriptions
