+++ src/core/game/GameMapLoader.ts
@@ -1,0 +1,13 @@
+import { GameMapType } from "./Game";
+import { MapManifest } from "./TerrainMapLoader";
+
+export interface GameMapLoader {
+  getMapData(map: GameMapType): MapData;
+}
+
+export interface MapData {
+  mapBin: () => Promise<Uint8Array>;
+  miniMapBin: () => Promise<Uint8Array>;
+  manifest: () => Promise<MapManifest>;
+  webpPath: () => Promise<string>;
+}
